requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
Django==4.1.3
django-ckeditor==6.5.1
django-js-asset==2.0.0
djlint==1.19.7
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
importlib-metadata==5.0.0
jsbeautifier==1.14.7
pathspec==0.10.2
Pillow==9.3.0
psycopg2-binary==2.9.5
PyYAML==6.0
regex==2022.10.31
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
tzdata==2022.6
zipp==3.10.0
